backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.21.0
mdurl==0.1.2
oauthlib==3.3.1
openai==2.7.1
openai-agents==0.5.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.28.0
rsa==4.9.1
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.49.3
tqdm==4.67.1
typer==0.20.0
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
tzlocal==5.2
